backend/pyproject.toml
[project]
name = "multi-agent-research-assistant"
version = "0.1.0"
description = "TODO-driven multi-agent deep research assistant (FastAPI + LangGraph)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "httpx>=0.27.0",
  "chromadb>=0.5.0",
  "langchain-core>=0.2.20",
  "langchain-openai>=0.1.20",
  "langgraph>=0.2.0",
  "markdown-it-py>=3.0.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "pypdf>=4.2.0",
  "python-multipart>=0.0.9",
  "rank-bm25>=0.2.2",
  "duckduckgo-search>=6.3.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]
local_embedding = [
  "sentence-transformers>=3.0.0",
]
ocr = [
  "pillow>=10.4.0",
  "pytesseract>=0.3.10",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
